Fix data type

+++ 006_hw_functions/taskC6HW_grainsOnBoard.c
@@ -1,10 +1,12 @@
+#include <stdio.h>
+
 // How many grains are on the board
-int grainsOnBoard (int num) {
-    int sum = 1;
+unsigned long long grainsOnBoard (int num) {
+    unsigned long long sum = 1;
 
     for (int i = 1; i < num; i++) {
         sum *= 2;
-        printf ("%d\n", sum);
+        printf ("%llu\n", sum);
     }
 
     return sum;
@@ -13,7 +15,7 @@ int grainsOnBoard (int num) {
 int main() {
     int num;
     scanf ("%d", &num);
-    printf ("\n%d", grainsOnBoard(num));
+    printf ("\n%llu", grainsOnBoard(num));
 
     return 0;
 }
